feat(#47) : 게시글 유저 기능 추가

+++ src/main/java/com/hobbyhop/domain/post/service/impl/PostServiceImpl.java
@@ -1,11 +1,20 @@
 package com.hobbyhop.domain.post.service.impl;
 
+import com.hobbyhop.domain.club.entity.Club;
+import com.hobbyhop.domain.club.repository.ClubRepository;
+import com.hobbyhop.domain.club.service.ClubService;
 import com.hobbyhop.domain.post.dto.PostRequestDTO;
 import com.hobbyhop.domain.post.dto.PostResponseDTO;
 import com.hobbyhop.domain.post.entity.Post;
 import com.hobbyhop.domain.post.repository.PostRepository;
 import com.hobbyhop.domain.post.service.PostService;
+import com.hobbyhop.domain.user.entity.User;
+import com.hobbyhop.domain.user.repository.UserRepository;
+import com.hobbyhop.domain.user.service.UserService;
+import com.hobbyhop.global.exception.club.ClubNotFoundException;
+import com.hobbyhop.global.exception.post.PostNotCorrespondUser;
 import com.hobbyhop.global.exception.post.PostNotFoundException;
+import com.hobbyhop.global.security.userdetails.UserDetailsImpl;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -16,7 +25,9 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class PostServiceImpl implements PostService {
 
+    private final UserRepository userRepository;
     private final PostRepository postRepository;
+    private final ClubRepository clubRepository;
 
     @Override
     public Post findById(Long postId) {
@@ -26,11 +37,15 @@ public class PostServiceImpl implements PostService {
 
     @Override
     @Transactional
-    public PostResponseDTO makePost(Long clubId, PostRequestDTO postRequestDTO) {
+    public PostResponseDTO makePost(UserDetailsImpl userDetails, Long clubId, PostRequestDTO postRequestDTO) {
+
+        Club club = clubRepository.findById(clubId).orElseThrow(ClubNotFoundException::new);
 
         Post post = Post.builder()
                 .postTitle(postRequestDTO.getPostTitle())
                 .postContent(postRequestDTO.getPostContent())
+                .club(club)
+                .user(userDetails.getUser())
                 .build();
 
         Post savedPost = postRepository.save(post);
@@ -64,9 +79,15 @@ public class PostServiceImpl implements PostService {
 
     @Override
     @Transactional
-    public PostResponseDTO modifyPost(Long clubId, Long postId, PostRequestDTO postRequestDTO){
+    public PostResponseDTO modifyPost(UserDetailsImpl userDetails, Long clubId, Long postId, PostRequestDTO postRequestDTO){
 
         Post post = postRepository.findById(postId).orElseThrow(PostNotFoundException::new);
+
+        User dbuser = userRepository.findById(userDetails.getUser().getId()).orElseThrow();
+
+        if(!dbuser.getId().equals(post.getUser().getId())){
+            throw new PostNotCorrespondUser();
+        }
 
         if(postRequestDTO.getPostTitle() != null) {
             post.changeTitle(postRequestDTO.getPostTitle());
@@ -87,7 +108,15 @@ public class PostServiceImpl implements PostService {
 
     @Override
     @Transactional
-    public void deletePost(Long clubId, Long postId){
+    public void deletePost(UserDetailsImpl userDetails, Long clubId, Long postId){
+
+        Post post = postRepository.findById(postId).orElseThrow(PostNotFoundException::new);
+
+        User dbuser = userRepository.findById(userDetails.getUser().getId()).orElseThrow();
+
+        if(!dbuser.getId().equals(post.getUser().getId())){
+            throw new PostNotCorrespondUser();
+        }
 
         postRepository.deleteById(postId);
     }
